# src/docs/cooking-api.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cooking API",
    "description": "Documentation for Cooking API",
    "contact": {
      "email": "e.garciadececa@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API"
    }
  ],
  "paths": {
    "/ingredients": {
      "get": {
        "tags": [
          "ingredients"
        ],
        "description": "Get all available ingredients.",
        "responses": {
          "200": {
            "description": "A list of all ingredients.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "ingredients"
        ],
        "description": "Creates an ingredient",
        "requestBody": {
          "description": "Ingredient to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created ingredient.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/ingredients/{id}": {
      "get": {
        "tags": [
          "ingredients"
        ],
        "description": "Get an ingredient by id.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested ingredient.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "ingredients"
        ],
        "description": "Updates an ingredient",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "description": "Ingredient data to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated ingredient.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "ingredients"
        ],
        "description": "Deletes an ingredient by id.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/ingredients/{id}/outcomes": {
      "get": {
        "tags": [
          "ingredients",
          "steps"
        ],
        "description": "Get all steps that use this ingredient as source.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the steps that use this ingredient as source.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DetailedStep"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/ingredients/{id}/sources": {
      "get": {
        "tags": [
          "ingredients",
          "steps"
        ],
        "description": "Get all steps that result into this ingredient.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the steps that result into this ingredient.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DetailedStep"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/recipes": {
      "get": {
        "tags": [
          "recipes"
        ],
        "description": "Get all available recipes.",
        "parameters": [
          {
            "$ref": "#/components/parameters/detailed"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all recipes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recipe"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/steps": {
      "get": {
        "tags": [
          "steps"
        ],
        "description": "Get all available steps.",
        "parameters": [
          {
            "$ref": "#/components/parameters/detailed"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all steps.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Step"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "steps"
        ],
        "description": "Creates an step",
        "requestBody": {
          "description": "Step to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleStep"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created ingredient.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedStep"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "steps"
        ],
        "description": "Deletes an step",
        "requestBody": {
          "description": "Step to delete",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleStep"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/utensils": {
      "get": {
        "tags": [
          "utensils"
        ],
        "description": "Get all avaliable utensils.",
        "responses": {
          "200": {
            "description": "A list of all utensils.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Utensil"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "utensils"
        ],
        "description": "Creates an utensil",
        "requestBody": {
          "description": "Utensil to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtensilData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created utensil.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Utensil"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/utensils/{id}": {
      "get": {
        "tags": [
          "utensils"
        ],
        "description": "Get an utensil by id.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested utensil.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Utensil"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "utensils"
        ],
        "description": "Updates an utensil",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "description": "Utensil data to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtensilData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated utensil.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Utensil"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "utensils"
        ],
        "description": "Deletes an utensil by id.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/utensils/{id}/uses": {
      "get": {
        "tags": [
          "utensils",
          "steps"
        ],
        "description": "Get all steps that use this utensil.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the steps that use this utensil.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DetailedStep"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "description": "ID to get",
        "required": true,
        "schema": {
          "type": "number"
        }
      },
      "detailed": {
        "name": "detailed",
        "in": "query",
        "description": "If detailed, all the info of the ingredients and utensils is returned. If not, only the ids are returned.",
        "required": false,
        "schema": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "responses": {
      "204": {
        "description": "No content. Operation was performed successfully and no data needs to be returned"
      },
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "More info about the error"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "More info about the error"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Generic error message"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Ingredient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "start",
              "mid",
              "end"
            ]
          }
        }
      },
      "Recipe": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/SimpleRecipe"
          },
          {
            "$ref": "#/components/schemas/DetailedRecipe"
          }
        ]
      },
      "Step": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/SimpleStep"
          },
          {
            "$ref": "#/components/schemas/DetailedStep"
          }
        ]
      },
      "Utensil": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "waitTimeInMillis": {
            "type": "integer"
          }
        }
      },
      "DetailedRecipe": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "input": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "utensil1": {
            "$ref": "#/components/schemas/Utensil"
          },
          "mid1": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "utensil2": {
            "$ref": "#/components/schemas/Utensil"
          },
          "mid2": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "utensil3": {
            "$ref": "#/components/schemas/Utensil"
          },
          "mid3": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "utensil4": {
            "$ref": "#/components/schemas/Utensil"
          },
          "mid4": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "utensil5": {
            "$ref": "#/components/schemas/Utensil"
          },
          "mid5": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "output": {
            "$ref": "#/components/schemas/Ingredient"
          }
        }
      },
      "SimpleRecipe": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "input": {
            "type": "integer"
          },
          "utensil1": {
            "type": "integer"
          },
          "mid1": {
            "type": "integer"
          },
          "utensil2": {
            "type": "integer"
          },
          "mid2": {
            "type": "integer"
          },
          "utensil3": {
            "type": "integer"
          },
          "mid3": {
            "type": "integer"
          },
          "utensil4": {
            "type": "integer"
          },
          "mid4": {
            "type": "integer"
          },
          "utensil5": {
            "type": "integer"
          },
          "mid5": {
            "type": "integer"
          },
          "output": {
            "type": "integer"
          }
        }
      },
      "DetailedStep": {
        "type": "object",
        "properties": {
          "input": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "utensil": {
            "$ref": "#/components/schemas/Utensil"
          },
          "output": {
            "$ref": "#/components/schemas/Ingredient"
          }
        }
      },
      "SimpleStep": {
        "type": "object",
        "properties": {
          "input": {
            "type": "integer"
          },
          "utensil": {
            "type": "integer"
          },
          "output": {
            "type": "integer"
          }
        }
      },
      "IngredientData": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "start",
              "mid",
              "end"
            ]
          }
        }
      },
      "UtensilData": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "waitTimeInMillis": {
            "type": "integer"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "ingredients",
      "description": "Methods to manage ingredients and its uses"
    },
    {
      "name": "recipes",
      "description": "Methods to query recipes"
    },
    {
      "name": "steps",
      "description": "Methods to query recipe steps"
    },
    {
      "name": "utensils",
      "description": "Methods to manage utensils and its uses"
    }
  ]
}